==> classification/data_creation/3_sentences_for_classifiers_webis.ipynb <==
import os
# nvidia-smi
os.environ['CUDA_VISIBLE_DEVICES'] = '3'

import os
import json
import pandas as pd
import numpy as np
from tqdm import tqdm
tqdm.pandas()
from flair.tokenization import SegtokSentenceSplitter
from flair.data import Sentence
from flair.models import SequenceTagger
from nltk.stem import WordNetLemmatizer
from nltk.stem import PorterStemmer
import warnings
warnings.filterwarnings('ignore')

import logging
logger = logging.getLogger('flair')
logger.setLevel(level=logging.ERROR)
fh = logging.StreamHandler()
logger.addHandler(fh)

PREFIX = "../data/"
PA_PATH = PREFIX + "sap2017-connotation-frames-power-agency/"
S_PATH = PREFIX + "kiesel2017-webis-simple-sentences-17/"

# read power_agency

power_agency = pd.read_csv(PA_PATH + "agency_power_prepro.csv", sep=',')

# general tools

splitter = SegtokSentenceSplitter()
lemmatizer = WordNetLemmatizer()
stemmer = PorterStemmer()
pos_tagger = SequenceTagger.load("upos-fast")

# get sentences

@InProceedings{kiesel:2017a,
  author =                {Johannes Kiesel and Benno Stein and Stefan Lucks},
  booktitle =             {24th Annual Network and Distributed System Security Symposium (NDSS 2017)},
  doi =                   {10.14722/ndss.2017.23077},
  ids =                   {stein:2017a},
  month =                 feb,
  numpages =              13,
  publisher =             {Association for Computational Linguistics},
  site =                  {San Diego, CA, USA},
  title =                 {{A Large-scale Analysis of the Mnemonic Password Advice}},
  year =                  2017
}

Get data at https://zenodo.org/record/205950#.Yi8qtHrMJhE

test = pd.read_csv(S_PATH + 'webis-simple-sentences-17-corpus-test.txt', delimiter = "\n", names=['text'])

test.head()

len(test)

wikisents = pd.read_pickle(PA_PATH + 'power_agency_wikisents.pkl')

jokesents = pd.read_pickle(PA_PATH + 'power_agency_jokesents.pkl')

sents = jokesents.merge(wikisents)

sents['sents'] = sents['jsents'] + sents['wsents']
sents = sents.drop(['jsents','wsents'],axis=1)

# determine missing verbs

## filter too long sentences

sents.sents.apply(lambda x: sum([len(y) for y in x])/len(x) if len(x) > 0 else 0).describe()

sents.sents.apply(lambda x: sum([len(y) for y in x])).sum()

def filter_long_sents(row):
    return [s for s in row if len(s) <= 50]        

sents['sents'] = sents.sents.progress_apply(filter_long_sents)

sents.sents.apply(lambda x: sum([len(y) for y in x])).sum()

sents.sents.apply(len).describe()

## filter dataset

goal_n = 250

sents['missing'] = sents.sents.apply(lambda x: goal_n - len(x) if len(x) < goal_n else 0)

sum(sents['missing'])

### all test.texts to Sentences

test = test[test.text.apply(type) == str] # three were NaN

test = test[test.text.apply(len) >= 5]

len(test)

test['sent'] = test.text.apply(Sentence)

### filter by stems

sents['stem'] = sents.verb.progress_apply(stemmer.stem)

sents[sents.missing > 0]

def filter_long_test_sents(row):
    if len(row) <= 50:
        return row
    else:
        None 

test['sent'] = test.sent.progress_apply(filter_long_test_sents)

missing_stems = set(sents[sents.missing > 0].stem)

def pre_filter_sentences(sent):
    if sent and len(sent)>0:
        for token in sent:
            for stem in missing_stems:
                if stem in token.text:
                    return True
    return False

test['candidate'] = test['sent'].progress_apply(pre_filter_sentences)

test['candidate'] = test['candidate'].apply(lambda x: False if type(x) == float else x) # nan -> False

candidates = test[test['candidate']]

print('# Candidates:', len(candidates))

# get sentences

pa_verbprep = power_agency[power_agency.prep.notna()][['lemma', 'prep']]
pa_verbprep = list(pa_verbprep.itertuples(index=False, name=None))

pa_verblemma = set(power_agency.lemma)

def preprocess_webis(sent, col):
    if len(sent) > 0:
        pos_tagger.predict(sent)
        for token in sent:
            if token.get_tag('pos').value == 'VERB':
                token.add_tag('lemma', lemmatizer.lemmatize(token.text, pos='v'))
             
        for tid in range(len(sent)):
            token = sent[tid]
            if token.get_tag('pos').value == 'VERB' and token.get_tag('lemma').value in pa_verblemma:
                # verb from pa -> has prep?
                if tid+1 < len(sent):
                    next_token = sent[tid+1]
                    
                    if (token.get_tag('lemma').value, next_token.text) in pa_verbprep:
                        # has prep
                        index = power_agency[(power_agency.lemma == token.get_tag('lemma').value) & (power_agency.prep == next_token.text)].index
                    else:
                        index = power_agency[(power_agency.lemma == token.get_tag('lemma').value) & (~power_agency.prep.notna())].index
                else:
                    index = power_agency[(power_agency.lemma == token.get_tag('lemma').value) & (~power_agency.prep.notna())].index
                    
                # add to power_agency    
                if len(index) == 1:
                    result = power_agency.loc[index[0], col]
                    
                    if len(result) < upper_bound:
                        result.append(sent)
                        power_agency.at[index[0], col] = result
                else:
                    if len(index) > 1:
                        print(sent)
                        print(index)

upper_bound = 250

col = 'ssents'

power_agency[col] = [list() for x in range(len(power_agency.index))]

candidates['test'] = candidates['sent'].progress_apply(lambda s: preprocess_webis(s, col)) # run by script

power_agency.to_pickle(PA_PATH + 'power_agency_webissents.pkl')
