==> classification/data_creation/4_prepare_sentences.ipynb <==
import os
# nvidia-smi
os.environ['CUDA_VISIBLE_DEVICES'] = '4'

import os
import json
import random
import pandas as pd
from tqdm import tqdm
tqdm.pandas()
from flair.tokenization import SegtokSentenceSplitter
from flair.data import Sentence
from flair.models import SequenceTagger
from nltk.stem import WordNetLemmatizer
from flair.embeddings import FlairEmbeddings, StackedEmbeddings, ELMoEmbeddings
import warnings
warnings.filterwarnings('ignore')

import logging
logger = logging.getLogger('flair')
logger.setLevel(level=logging.ERROR)
fh = logging.StreamHandler()
logger.addHandler(fh)

PREFIX = "../data/"
PA_PATH = PREFIX + "sap2017-connotation-frames-power-agency/"
J_PATH = PREFIX + "pungas2017-plaintext-jokes/"
W_PATH = PREFIX + "wang2018-wiki-dataset/"
S_PATH = PREFIX + "kiesel2017-webis-simple-sentences-17/"

# read power_agency

wikisents = pd.read_pickle(PA_PATH + 'power_agency_wikisents.pkl')

jokesents = pd.read_pickle(PA_PATH + 'power_agency_jokesents.pkl')

webissents = pd.read_pickle(PA_PATH + 'power_agency_webissents.pkl')

sents = jokesents.merge(wikisents)
sents = sents.merge(webissents)
del wikisents, jokesents, webissents

sents['sents'] = sents['jsents'] + sents['wsents']
sents['sents'] = sents['sents'] + sents['ssents']

sum(sents.jsents.apply(len)), sum(sents.wsents.apply(len)), sum(sents.ssents.apply(len)), (sum(sents.jsents.apply(len)) + sum(sents.wsents.apply(len)) + sum(sents.ssents.apply(len)))

sents = sents.drop(['jsents','wsents','ssents'],axis=1)

sum(sents.sents.apply(len))

len(sents)

sents.sents.apply(len).describe()

# filter long sentences

s = sents.explode('sents')

s.describe()

s.sents.apply(len).describe()

s = s[s.sents.apply(len) <= 75]

s.sents.apply(len).describe()

del s

sum(sents.sents.apply(len))

def filter_long_sents(row):
    return [s for s in row if len(s) <= 75]        

sents['sents'] = sents.sents.progress_apply(filter_long_sents)

sum(sents.sents.apply(len))

# balance dataset

sents.sents.apply(len).describe()

for i in range(0, 251, 5):
    print(i, ':', len(sents[sents.sents.apply(len) < i]), len(sents[sents.sents.apply(len) >= i]))

48 / 2144

n = 200

# filter verbs with too few sentences
sents = sents[sents.sents.apply(len) >= n]

# remove sentences that are too many
def filter_too_many(slist):
    return random.sample(slist, n)

sents['sents'] = sents.sents.progress_apply(filter_too_many)

sents = sents.reset_index()

sents.sents.apply(len).describe()

sents.to_pickle(PA_PATH + 'power_agency_sents.pkl')
