==> classification/data_creation/5_train_test_split.ipynb <==
import os
# nvidia-smi
os.environ['CUDA_VISIBLE_DEVICES'] = '2'

import pandas as pd
from tqdm import tqdm
tqdm.pandas()
import pickle

PREFIX = "../data/"
PA_PATH = PREFIX + "sap2017-connotation-frames-power-agency/"

# read data

pa = pd.read_pickle(PA_PATH + 'power_agency_sents.pkl')

pa = pa.drop('index',axis=1)

pa.head()

pa.verb.is_unique

pa.verb_prep.is_unique

len(pa)

# agency

agency = pa[pa.agency.notna()]

len(agency)

len(agency) * 0.2

419 / len(agency)

test_agency = agency.sample(419)

test_agency.head()

test_agency.describe()

test_agency_vp = set(test_agency.verb_prep)

test_agency_idx = set(test_agency.index)

len(test_agency_idx), len(test_agency_vp)

train_agency_vp = set(agency[~agency.verb_prep.isin(test_agency_vp)].verb_prep)

train_agency_idx = set(agency[~agency.verb_prep.isin(test_agency_vp)].index)

len(train_agency_vp), len(train_agency_idx)

## save

with open(PA_PATH + 'train_agency_vp.pickle', 'wb') as f:
    pickle.dump(train_agency_vp, f)

with open(PA_PATH + 'train_agency_idx.pickle', 'wb') as f:
    pickle.dump(train_agency_idx, f)

with open(PA_PATH + 'test_agency_vp.pickle', 'wb') as f:
    pickle.dump(test_agency_vp, f)

with open(PA_PATH + 'test_agency_idx.pickle', 'wb') as f:
    pickle.dump(test_agency_idx, f)

# power

power = pa[pa.power.notna()]

len(power)

len(power) * 0.2

340 / len(power)

test_power = power.sample(340)

test_power.head()

test_power.describe()

test_power_vp = set(test_power.verb_prep)

test_power_idx = set(test_power.index)

len(test_power_idx), len(test_power_vp)

train_power_vp = set(power[~power.verb_prep.isin(test_power_vp)].verb_prep)

train_power_idx = set(power[~power.verb_prep.isin(test_power_vp)].index)

len(train_power_vp), len(train_power_idx)

## save

with open(PA_PATH + 'train_power_vp.pickle', 'wb') as f:
    pickle.dump(train_power_vp, f)

with open(PA_PATH + 'train_power_idx.pickle', 'wb') as f:
    pickle.dump(train_power_idx, f)

with open(PA_PATH + 'test_power_vp.pickle', 'wb') as f:
    pickle.dump(test_power_vp, f)

with open(PA_PATH + 'test_power_idx.pickle', 'wb') as f:
    pickle.dump(test_power_idx, f)
